Implement SOP strategy logic and technician-based proposal filtering
- **Backend (C++):**
    - **ASRH (Agent Specialist in Repair Handling):**
        - Implement "Early" vs. "Late" maintenance start calculations to support the **SOP (Schedule Orientation Policy)** strategy.
        - Add logic to select the latest safe start time when the SOP strategy is active, pushing maintenance toward the deadline to maximize resource availability.
        - Default to the earliest possible start (SOM strategy) if SOP is inactive or the late option is deemed unsafe due to TBM (Time-Based Maintenance) conflicts.
        - Refactor and streamline constraint validation logic (C1–C5) for better performance and readability.
    - **AMS (Agent Machine Specialist):**
        - Introduce a "Smart Filter" in `handleAnomaly` that groups proposals by technician (`arhId`) and retains only the optimal proposal (lowest $f$ score) for each.
        - Update the scheduling flow to evaluate all raw physical options before selecting the unique best-per-ARH set for the UI.
        - Refactor multi-machine negotiation to use the filtered `bestOverall` proposal.
    - **Logging:**
        - Consolidate `jsonLog` calls into concise, single-line outputs across `ASRH.h` and `AMS.h` to reduce standard output overhead.
        - Improve clarity of strategy selection logs (e.g., explicitly stating if SOP or SOM start times were chosen).

- **Build:**
    - Update `core_engine.exe` binary.

+++ composeApp/src/jvmMain/kotlin/com/example/smartfactorymas/SimulationDomain.kt
@@ -261,8 +261,8 @@ data class GanttState(
     val masOutput: MASOutput? = null,
     val multiMachineOutput: MultiMachineResultEvent? = null,
     val logs: List<LogEvent> = emptyList(),
-    val f1: Double = 0.0,
-    val f2: Double = 0.0,
-    val f: Double = 0.0,
+    val f1: Float = 0.0f,
+    val f2: Float = 0.0f,
+    val f: Float = 0.0f,
     val chosenArh: String = ""
 )
